# logging subsystem rewrite