--- a/lib/clippy.c
+++ b/lib/clippy.c
@@ -107,21 +107,11 @@ int main(int argc, char **argv)
 #include "log.h"
 #include "zassert.h"
 
-#define ZLOG_FUNC(FUNCNAME)                                                    \
-	void FUNCNAME(const char *format, ...)                                 \
-	{                                                                      \
-		va_list args;                                                  \
-		va_start(args, format);                                        \
-		vfprintf(stderr, format, args);                                \
-		fputs("\n", stderr);                                           \
-		va_end(args);                                                  \
-	}
-
-ZLOG_FUNC(zlog_err)
-ZLOG_FUNC(zlog_warn)
-ZLOG_FUNC(zlog_info)
-ZLOG_FUNC(zlog_notice)
-ZLOG_FUNC(zlog_debug)
+void vzlog(int prio, const char *format, va_list args)
+{
+	vfprintf(stderr, format, args);
+	fputs("\n", stderr);
+}
 
 void _zlog_assert_failed(const char *assertion, const char *file,
 			 unsigned int line, const char *function)