--- a/lib/grammar_sandbox_main.c
+++ b/lib/grammar_sandbox_main.c
@@ -45,11 +45,7 @@ int main(int argc, char **argv)
 
 	master = thread_master_create(NULL);
 
-	openzlog("grammar_sandbox", "NONE", 0, LOG_CONS | LOG_NDELAY | LOG_PID,
-		 LOG_DAEMON);
-	zlog_set_level(ZLOG_DEST_SYSLOG, ZLOG_DISABLED);
-	zlog_set_level(ZLOG_DEST_STDOUT, LOG_DEBUG);
-	zlog_set_level(ZLOG_DEST_MONITOR, ZLOG_DISABLED);
+	zlog_aux_init("NONE: ", LOG_DEBUG);
 
 	/* Library inits. */
 	cmd_init(1);