--- a/lib/hook.c
+++ b/lib/hook.c
@@ -18,16 +18,26 @@
 #include "config.h"
 #endif
 
+#include <string.h>
+
 #include "memory.h"
 #include "hook.h"
 
 DEFINE_MTYPE_STATIC(LIB, HOOK_ENTRY, "Hook entry")
 
-void _hook_register(struct hook *hook, void *funcptr, void *arg, bool has_arg,
-		    struct frrmod_runtime *module, const char *funcname,
-		    int priority)
+void _hook_register(struct hook *hook, struct hookent *stackent, void *funcptr,
+		    void *arg, bool has_arg, struct frrmod_runtime *module,
+		    const char *funcname, int priority)
 {
-	struct hookent *he = XCALLOC(MTYPE_HOOK_ENTRY, sizeof(*he)), **pos;
+	struct hookent *he, **pos;
+
+	if (!stackent->ent_used)
+		he = stackent;
+	else {
+		he = XCALLOC(MTYPE_HOOK_ENTRY, sizeof(*he));
+		he->ent_on_heap = true;
+	}
+	he->ent_used = true;
 	he->hookfn = funcptr;
 	he->hookarg = arg;
 	he->has_arg = has_arg;
@@ -52,7 +62,10 @@ void _hook_unregister(struct hook *hook, void *funcptr, void *arg, bool has_arg)
 		if (he->hookfn == funcptr && he->hookarg == arg
 		    && he->has_arg == has_arg) {
 			*prev = he->next;
-			XFREE(MTYPE_HOOK_ENTRY, he);
+			if (he->ent_on_heap)
+				XFREE(MTYPE_HOOK_ENTRY, he);
+			else
+				memset(he, 0, sizeof(*he));
 			break;
 		}
 }