--- a/lib/subdir.am
+++ b/lib/subdir.am
@@ -45,6 +45,7 @@ lib_libfrr_la_SOURCES = \
 	lib/libfrr.c \
 	lib/linklist.c \
 	lib/log.c \
+	lib/log_filter.c \
 	lib/log_vty.c \
 	lib/md5.c \
 	lib/memory.c \