… doc/developer/logging.rst
… doc/user/basic.rst
… doc/user/setup.rst
… lib/log.c
… lib/log.h
… lib/log_filter.c
… lib/log_int.h
… lib/log_vty.c
… lib/memory.c
… lib/memory.h
… lib/thread.c
… lib/zlog.c
… lib/zlog.h
… lib/zlog_targets.c
… lib/zlog_targets.h
… pimd/pim_mlag.c
… sharpd/sharp_logpump.c
… tests/bgpd/test_peer_attr.c
… tests/lib/cli/common_cli.c
… tests/lib/northbound/test_oper_data.c
… tests/lib/test_segv.c
… tests/lib/test_sig.c
… tests/lib/test_zlog.c
… vtysh/vtysh.c
… watchfrr/watchfrr.c
… watchfrr/watchfrr_vty.c