This is a full rewrite of the logging pieces in `libfrr`;  the old code was... old.  The new code uses a lock-free linked list + RCU to maintain the log targets, meaning it's possible for the entire logging system to be lock free.  (Individual logging targets may still need a lock, but none of file/fd/syslog have one.)

Note that some functionality is lost in this patch:
- Solaris printstack() backtraces are ditched (unlikely to come back)
- ~the `log-filter` machinery is gone (also unlikely to come back... I have a better mechanism queued up)~ re-added.
- `terminal monitor` is temporarily stubbed out.  The old code had a race condition with VTYs going away.  It'll likely come back rewritten and with vtysh support.
- The `zebra_ext_log` hook is gone.  Instead, it's now much easier to add a "proper" logging target.

I have further logging infrastructure pending which builds on top of this, this is just the first chunk.

The new logging code is permissively licensed like all other `lib` pieces that I did a full rewrite on.