@eqvinox 2019-12-02
@louberger need info on failure here. As a guess I'd say you need to remove an `#include "log_int.g"` somewhere.