@louberger 2019-12-03
is there a replacement for vzlog_test?