@louberger 2019-12-04
looks like there's a leak/unfreed object:

==32594==    by 0x27FF8F: qcalloc (memory.c:110)
==32594==    by 0x2C3333: zlog_target_clone (zlog.c:231)
==32594==    by 0x2C3903: zlog_file_cycle (zlog_targets.c:179)
==32594==    by 0x2C3A8D: zlog_file_set_other (zlog_targets.c:203)
==32594==    by 0x27E84F: config_log_stdout_magic (log_vty.c:207)
==32594==    by 0x27E84F: config_log_stdout (log_vty_clippy.c:42)
==32594==    by 0x25E2AD: cmd_execute_command_real.isra.2 (command.c:964)
==32594==    by 0x2604F6: cmd_execute_command_strict (command.c:1072)
==32594==    by 0x2604F6: command_config_read_one_line (command.c:1221)
==32594==    by 0x26075C: config_from_file (command.c:1281)
==32594==    by 0x2B62AD: vty_read_file (vty.c:2297)
==32594==    by 0x2B62AD: vty_read_config (vty.c:2513)
==32594==    by 0x27A6B6: frr_config_read_in (libfrr.c:853)
==32594==    by 0x2AD7BB: thread_call (thread.c:1549)
==32594== 
{
   <insert_a_suppression_name_here>
   Memcheck:Leak
   match-leak-kinds: reachable
   fun:calloc
   fun:qcalloc
   fun:zlog_target_clone
   fun:zlog_file_cycle
   fun:zlog_file_set_other
   fun:config_log_stdout_magic
   fun:config_log_stdout
   fun:cmd_execute_command_real.isra.2
   fun:cmd_execute_command_strict
   fun:command_config_read_one_line
   fun:config_from_file
   fun:vty_read_file
   fun:vty_read_config
   fun:frr_config_read_in
   fun:thread_call
}
==32594== LEAK SUMMARY:
==32594==    definitely lost: 0 bytes in 0 blocks
==32594==    indirectly lost: 0 bytes in 0 blocks
==32594==      possibly lost: 0 bytes in 0 blocks
==32594==    still reachable: 88 bytes in 1 blocks
==32594==         suppressed: 32 bytes in 1 blocks
==32594==