@eqvinox 2019-12-05
@louberger 
> is there a replacement for `vzlog_test`?

No; there is considerably more fine-grained control over logging output coming in follow-up PRs which makes it semantically ill-defined to give a generic answer to "is this log level enabled?"