@eqvinox 2019-12-05
> looks like there's a leak/unfreed object:

This is intentional, please add a valgrind suppression.  The builtin memtype leak tracking has `DECLARE_MGROUP_ACTIVEATEXIT` for this but that's not something valgrind understands.