@eqvinox 2019-12-05
Also, to extend on `vzlog_test` — wrapping `zlog()`/`vzlog()` in functions like the rfapi code does will result in some of the new functionality being lost/unavailable for any wrapped logging calls.  (No unique ID, no code location.)