@louberger 2019-12-17
@eqvinox failures are memory free on exit related:

7.28.1    7.0.0.130 Memory leak on exit, 1 errors                  0      1
7.28.2    7.0.0.130 Leaked type: log file target : 1 * 88 blocks   0      1