@eqvinox 2019-12-17
@louberger these are intentional, and the exit status from `debug memstats-at-exit` reflects that correctly (0 exit status).  I've also added them to the valgrind suppression file.  That's all I can do / have access to, anything else you need to fix on your end.

```
7.28.1 7.0.0.130 Memory leak on exit, 1 errors 0 1
7.28.2 7.0.0.130 Leaked type: log file target : 1 * 88 blocks 0 1
```