@eqvinox 2020-01-16
@louberger it now prints `(active at exit)` at the end of the line, e.g.
```
bgpd: showing active allocations in memory group libfrr
bgpd: memstats:  Hook entry                    :      4 *         48
bgpd: memstats:  Prefix                        :      3 *         48
bgpd: showing active allocations in memory group logging subsystem
bgpd: memstats:  log file target               :      1 *         88 (active at exit)
bgpd: showing active allocations in memory group bgpd
bgpd: memstats:  BGP connected                 :      1 *          4
bgpd: showing active allocations in memory group rfapi
```