@louberger 2020-01-16
thanks for this. if the number is unexpected, e.g., 2*nn, is the line 
still printed?  If not, how does someone know if there's a leak?

On 1/16/2020 3:32 PM, David Lamparter wrote:
>
> @louberger <https://github.com/louberger> it now prints |(active at 
> exit)| at the end of the line, e.g.
>
> |bgpd: showing active allocations in memory group libfrr bgpd: 
> memstats: Hook entry : 4 * 48 bgpd: memstats: Prefix : 3 * 48 bgpd: 
> showing active allocations in memory group logging subsystem bgpd: 
> memstats: log file target : 1 * 88 (active at exit) bgpd: showing 
> active allocations in memory group bgpd bgpd: memstats: BGP connected 
> : 1 * 4 bgpd: showing active allocations in memory group rfapi |
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub 
> <https://github.com/FRRouting/frr/pull/5451?email_source=notifications&email_token=ABSFNSFQ6RGXL7ILWI5ZDNLQ6C75BA5CNFSM4JTYTM2KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEJFOJJY#issuecomment-575333543>, 
> or unsubscribe 
> <https://github.com/notifications/unsubscribe-auth/ABSFNSGCU5ZE7VKJR55K3TTQ6C75BANCNFSM4JTYTM2A>.
>