@eqvinox 2020-01-16
> thanks for this. if the number is unexpected, e.g., 2*nn, is the line
still printed?  If not, how does someone know if there's a leak?

Tracking what number is expected isn't really possible, it depends on config statements which aren't available to the backend code (there's 2 layer breaks inbetween that.)

If your config stays the same, the numbers should stay the same, so the best check would be to compare it against reference output.  (If your logging config is the same for all tests, the reference can also be the same.)

Also need to fix the `Hook entry` type to use this too, those 4 items in there are also logging related.