@eqvinox 2020-01-16
(Actually the `Hook entry` MTYPE I can maybe fix, now that I think about it...  will try and see...)