@eqvinox 2020-01-17
quick preliminary perf numbers:

before:
```
  2291000 messages written
 25000000 messages expected =   9.2%
     1.908609 usr     833.1ns/msg
     2.948940 sys    1287.2ns/msg
```
after:
```
  4091000 messages written
 25000000 messages expected =  16.4%
     3.344797 usr     817.6ns/msg
     1.612384 sys     394.1ns/msg
```

Done without controlling for external factors on my laptop, so take with a grain of salt; I'll grab controlled numbers in a bit.

(That said it matches my expectations; userspace cost is pretty much the same but syscalls/kernel cost is much lower through batching.)