@louberger 2020-01-21
Do you have an update on this?

On 1/16/2020 4:15 PM, David Lamparter wrote:
>
> (Actually the |Hook entry| MTYPE I can maybe fix, now that I think 
> about it... will try and see...)
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub 
> <https://github.com/FRRouting/frr/pull/5451?email_source=notifications&email_token=ABSFNSDFTCQU4NSFZEQGBYDQ6DE5TA5CNFSM4JTYTM2KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEJFSJZA#issuecomment-575349988>, 
> or unsubscribe 
> <https://github.com/notifications/unsubscribe-auth/ABSFNSB2JFUT2R5TKJE4RZ3Q6DE5TANCNFSM4JTYTM2A>.
>