@eqvinox 2020-01-21
@louberger I thought you're updating your CI to look at the `(active at exit)` marker?

I did fix the `Hook entry` leak so that won't be getting the marker, but there's still a few core logging structures that need to stay alive until exit.